Controllers with transformer

+++ src/main/java/generalPackage/data/dao/NotebookDAO.java
@@ -3,5 +3,5 @@ package generalPackage.data.dao;
 import generalPackage.data.entity.Notebook;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface NotebookDAO extends JpaRepository<Notebook, Long> {
+public interface NotebookDAO extends JpaRepository<Notebook, Integer> {
 }
